% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elhorst.R
\name{elhorst}
\alias{elhorst}
\title{Elhorst procedure}
\usage{
elhorst(base, modele, matrice, alpha)
}
\arguments{
\item{base}{data}

\item{modele}{formula}

\item{matrice}{weights matrix}

\item{alpha}{confidence level between 0 and 1}
}
\value{
returns a dataframe of results
}
\description{
Elhorst procedure
}
