% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesage_pace.R
\name{lesage_pace}
\alias{lesage_pace}
\title{LeSage and Pace procedure}
\usage{
lesage_pace(base, modele, matrice, alpha)
}
\arguments{
\item{base}{data}

\item{modele}{formula}

\item{matrice}{weights matrix}

\item{alpha}{confidence level between 0 and 1}
}
\value{
returns a dataframe of results
}
\description{
LeSage and Pace procedure
}
